requirements.txt
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
hydra-core==1.3.2
omegaconf==2.3.0
wandb==0.16.0
albumentations==1.3.1
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.26.4
matplotlib==3.8.0
seaborn==0.13.0
gradio==4.0.0
pytest==7.4.0
torchmetrics==1.2.0
pillow==10.0.0
tqdm==4.66.0
opencv-python-headless==4.8.1.78

